applying more suggestions
